requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
transformers>=4.45.0
accelerate>=0.33.0
bitsandbytes>=0.43.0
peft>=0.12.0
datasets>=2.19.0
sentencepiece>=0.2.0
faiss-cpu>=1.8.0
langchain>=0.2.10
langchain-community>=0.2.10
langchain-text-splitters>=0.2.2
torch>=2.1.0
numpy>=1.26.0
pymupdf
trafilatura
duckduckgo-search
streamlit
requests>=2.31.0
rank-bm25>=0.2.2
sentence-transformers>=2.2.0
llama-cpp-python>=0.2.40
